+++ rust/operator-binary/src/controller.rs
@@ -378,10 +378,7 @@ impl OpaClusterConfigFile {
             decision_logs: decision_logging,
             // Enable more Prometheus metrics, such as bundle loads
             // See https://www.openpolicyagent.org/docs/monitoring#status-metrics
-            status: Some(OpaClusterConfigStatus {
-                service: OPA_STACKABLE_SERVICE_NAME.to_owned(),
-                prometheus: true,
-            }),
+            status: Some(OpaClusterConfigStatus { prometheus: true }),
         }
     }
 }
@@ -418,7 +415,6 @@ pub struct OpaClusterConfigDecisionLog {
 
 #[derive(Serialize, Deserialize)]
 struct OpaClusterConfigStatus {
-    service: String,
     prometheus: bool,
 }
 
@@ -448,6 +444,7 @@ pub async fn reconcile_opa(
         OPA_CONTROLLER_NAME,
         &opa.object_ref(&()),
         ClusterResourceApplyStrategy::from(&opa.spec.cluster_operation),
+        &opa.spec.object_overrides,
     )
     .context(FailedToCreateClusterResourcesSnafu)?;